backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1
google-genai==1.7.0
pymupdf==1.25.3
Pillow==11.1.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
